working on refactoring explore env
