+++ configs/dapo_config.yaml
@@ -2,7 +2,7 @@
 seed: 117
 
 logging:
-  output_dir: "./runs/test_dapo_grpo_enhanced_qwen2.5_0.5b_gsm8k"  # where to save all artifacts (tensorboard events, samples, checkpoints)
+  output_dir: "./runs/test_dapo_qwen2.5_0.5b_gsm8k"  # where to save all artifacts (tensorboard events, samples, checkpoints)
   enable_tensorboard: true
   sample_buffer_size: 100               # how often to save samples to external files
   sample_file_format: 'jsonl'           # external file type, support 'jsonl', 'jsonl.gz', 'parquet'
@@ -32,13 +32,13 @@ dapo_grpo:
   train_batch_size: 128
   gamma: 1.0                            # discount
   # DAPO improvements
-  clip_eps_high: 0.25                   # PPO-style policy clip higher bound
-  clip_eps_low: 0.15                    # PPO-style policy clip lower bound
-  length_max: 4096                      # Maximum completion length for length penalty reward
-  length_cache: 512                     # Maximum completion length cache buffer for length penalty reward
+  clip_eps_high: 0.28                   # PPO-style policy clip higher bound
+  clip_eps_low: 0.2                     # PPO-style policy clip lower bound
+  length_max: 1024                      # Maximum completion length for length penalty reward
+  length_cache: 128                     # Maximum completion length cache buffer for length penalty reward
   normalize_advantages: false
   kl_loss_coef: 0.0
-  entropy_loss_coef: 0.01
+  entropy_loss_coef: 0.001
   sync_reference_interval: 0            # every N global steps
 
   # for train rollout
