+++ configs/ppo_config.yaml
@@ -51,9 +51,8 @@ ppo:
   sync_reference_interval: 0            # every N global steps
 
   # for sample generation
-  max_prompt_tokens: 512                # skip sample question that's too long
   group_size: 16                         # for each sample we generate a group of answers
-  max_completion_tokens: 800
+  max_completion_tokens: 1024
   temperature: 0.9                      
   repetition_penalty: 1.0
   top_p: 0.95
