+++ src/rl4llm/envs/__init__.py
@@ -1,4 +1,3 @@
+from .explore_env import ExploreSglMDPEnv
 from .hf_env import HfMDPEnv
 from .sgl_env import SglMDPEnv
-
-# from .explore_env import ExploreSglMDPEnv
