+++ src/rl4llm/envs/explore_env.py
@@ -1,3 +1,5 @@
+"""Implements custom MDP ENV for collect samples using SGLang inference server with a custom HTTP client"""
+
 import functools
 import logging
 import random
@@ -7,8 +9,9 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 import numpy as np
 import torch
 
+from rl4llm.core.base_env import BaseMDPEnv, ChatMessage
 from rl4llm.core.base_inference_client import InferenceClient
-from rl4llm.envs.sgl_env import EnvState, SglMDPEnv
+from rl4llm.envs.sgl_env import EnvState
 from rl4llm.generation.sgl_explore_procesor import SglExploreLogitProcessor
 
 logger = logging.getLogger(__name__)
@@ -17,10 +20,9 @@ logger = logging.getLogger(__name__)
 # --- Using SGLang inference client ---
 
 
-class ExploreSglMDPEnv(SglMDPEnv):
-    """An extension of the standard SglMDPEnv
-    where we apply some custom logits processor to the generation process
-    to encourage exploration."""
+class ExploreSglMDPEnv(BaseMDPEnv):
+    """Simple one-step MDP Environment where we apply some custom logits processor
+    to the generation process to encourage exploration."""
 
     def __init__(
         self,
@@ -29,11 +31,6 @@ class ExploreSglMDPEnv(SglMDPEnv):
         random_start_steps: int,
         random_start_top_k: int,
         random_start_skip_n: int,
-        replace_source_tokens: Optional[List[int]] = None,
-        replace_target_tokens: Optional[List[int]] = None,
-        replace_top_k: int = 1,
-        replace_max_count: int = 3,
-        replace_prob: float = 0.5,
         **kwargs,
     ):
 
@@ -47,10 +44,6 @@ class ExploreSglMDPEnv(SglMDPEnv):
         if any(p < 0 for p in group_top_p):
             raise ValueError('group_top_p values cannot be negative')
         assert len(group_top_p) == len(group_temperature) >= 1
-        if not isinstance(replace_prob, float) or not (
-            0.0 <= replace_prob < 1.0
-        ):
-            raise ValueError('replace_prob must be a float between (0.0, 1.0).')
 
         self.group_temperature = group_temperature
         self.group_top_p = group_top_p
@@ -58,58 +51,52 @@ class ExploreSglMDPEnv(SglMDPEnv):
         self.random_start_top_k = random_start_top_k
         self.random_start_skip_n = random_start_skip_n
 
-        self.replace_source_tokens = replace_source_tokens
-        self.replace_target_tokens = replace_target_tokens
-        self.replace_top_k = replace_top_k
-        self.replace_max_count = replace_max_count
-        self.replace_prob = replace_prob
-
     def _prepare_logits_processor(self, explore_prob: float) -> Optional[str]:
         """Creates the explore logits processor string if conditions are met."""
-        # This logic is self-contained setup, so keeping it separate is reasonable.
-        if (
-            explore_prob > 0
-            and random.random() < explore_prob
-            and (self.random_start_steps > 0 or self.replace_max_count > 0)
-        ):
+        if explore_prob > 0 and random.random() < explore_prob:
             explore_logit_processor = SglExploreLogitProcessor(
                 random_start_steps=self.random_start_steps,
                 random_start_top_k=self.random_start_top_k,
                 random_start_skip_n=self.random_start_skip_n,
-                replace_source_tokens=self.replace_source_tokens,
-                replace_target_tokens=self.replace_target_tokens,
-                replace_top_k=self.replace_top_k,
-                replace_max_count=self.replace_max_count,
             )
             return explore_logit_processor.to_str()
         return None
 
-    def _generate_completions(
+    @torch.inference_mode()
+    def _run_interaction_loop(
         self,
+        env_state: EnvState,
         llm: InferenceClient,
         sampling_params: Dict[str, Any],
-        state: EnvState,
         **kwargs: Optional[Dict[str, Any]],
-    ) -> Tuple[List[str], List[torch.Tensor], List[int]]:
+    ) -> EnvState:
         """
-        Generates completions using the LLM with probabilistic retry logic for incorrect items.
+        Performs a single generation step for all samples using the SampleState structure.
 
         Args:
-            llm: The custom inference client.
-            sampling_params: Dictionary of generation arguments (e.g., max_new_tokens).
-            state: The current EnvState containing prompt and ground_truth.
-            **kwargs: Additional arguments, including 'explore_probability'.
+            env_state: The starting state containing a list of SampleState objects.
+            llm: The language model inference client.
+            sampling_params: Configuration for generation.
+            **kwargs: Additional arguments (unused in default).
 
         Returns:
-            - completion_texts: List of final decoded completion strings.
-            - completion_tokens: List of final completion token tensors.
+            EnvState: The final state after one generation step, with updated SampleStates.
         """
+        logger.debug(
+            f"Rank {self.rank}: Running single-step interaction loop with SampleState design."
+        )
+
+        # 1. Prepare inputs for the LLM from the list of SampleStates
+        # Convert message histories to prompt strings
+        batch_prompts = self._convert_to_batch_prompts(env_state)
+
         explore_eps = kwargs.get('explore_epsilon', 0.0)
         logit_processor = self._prepare_logits_processor(explore_eps)
 
-        batch_size = state.input_ids.shape[0]
-        batched_sampling_params = []
+        batch_size = len(batch_prompts)
 
+        # Build sampling params for each sequence
+        batched_sampling_params = []
         for i in range(batch_size):
             sp = {
                 'temperature': float(self.group_temperature[i]),
@@ -121,16 +108,31 @@ class ExploreSglMDPEnv(SglMDPEnv):
                 'max_new_tokens': sampling_params.get('max_new_tokens', 4096),
                 'custom_params': {
                     'step': 0,
-                    'replace_prob': self.replace_prob,
-                    'replace_count': 0,
                 },
             }
             batched_sampling_params.append(sp)
 
-        output = llm.generate(
-            prompts=state.prompt,
+        # 2. Call the inference API for LLM generation
+        outputs = llm.generate(
+            prompts=batch_prompts,
             sampling_params=batched_sampling_params,
             custom_logit_processor=logit_processor,
         )
 
-        return self._process_llm_output(output)
+        # 3. Update each SampleState object *in place*
+        for i, sample_state in enumerate(env_state.sample_states):
+            # Fallback to dummy text to ensure code works
+            generated_text = (
+                outputs[i].get('text', 'I can not answer this question').strip()
+            )
+            # Append the new assistant message to the sample's history
+            sample_state.messages.append(
+                ChatMessage(role='assistant', content=generated_text)
+            )
+
+            # Mark this sample as done and record the step
+            sample_state.done = True
+            sample_state.current_step = 1
+
+        # 4. Return the *modified* EnvState object
+        return env_state
